Be more spec compliant by emitting a starting clear code.
This was removed from the code in 20dbfb6495f751451697585917d3b3550410661f,
which I don't think was really intentional.  It seems like most decoders are
fine without a starting clear code (in table should start cleared anyway,
right?), but it is probably better to try to follow the spec and send it.