GIF decoder, improve performance by more than 100x.
This is the simplest and biggest optimization, before even getting into more
details of the algorithm and the table, we can simple be a bit better about how
we manage writing out the index stream.  Preallocate the entire index buffer,
which is known because it should match the expected number of pixels.
This removes the incremental building of the index buffer (via concat()s).  This
was causing the runtime to spend 90% of the time in memmove() via ArrayConcat.

Performance of a test frame went from 800ms to around 6ms.