Implement interlaced decoding.
GifWriter still writes frames non-interlaced.  There are a few
different ways to approach interlacing, the performance to keep in
mind is the speed of de-interlacing and the impact on non-interlaced
frames.  This patch tries to strike a balance of not adding too much
additional overhead for non-interlaced frames, which is by far the
most common case.  The approach is to do the interlace scanline
jumping during the blit.  An alternative approach could be to
de-interlace the index stream, either in-place or with a copied
buffer, but this would be less performant for interlaced frames.

This was hard to measure, but the additional check at the end of each
scanline for handling interlaced frames adds a small but acceptable
performance overhead for non-interlaced frames, maybe a few percent.