Fix improper interlace scanstride calculation for sub-frames.
There was a missing * 4 for the framestride, necessary to take it
from pixels to bytes/indices.  `framestride` tracks the difference
between the frame width and the viewport, which means this bug only
manifests on interlaced sub-frames -- frames that just draw a portion
of the viewport.  These calculations were a bit messy, because
framestride = width - framewidth, meaning framestride + framewidth
is simply width, the entire viewport width.  However some
calculations were done with framestride + framewidth instead of
just using the viewport width.  Hopefully the clarity of the whole
convoluted pixel dance is slightly improved. hop skip scan