Range / value check for invalid block size (issue #24).
Reading out of bounds of the buffer will return undefined, which failed the
idenity zero check.  This now throws an exception on a value that is undefined
or less than zero.  The alternative option would be to just try to continue,
but this would probably just lead to needing more checks further down the line
(ex in the decompression routine which also has to handle block sizes).

Note: This is normally only ever going to be used with unsigned input buffers
so a value less than zero isn't possible, but it didn't seem to hurt to add the
check anyway.